refactor(cli): move for file organization
