+++ internal/ui/cli/main_init/input_utility.go
@@ -1,4 +1,4 @@
-package ui
+package main_init
 
 import (
 	"fmt"
